Stop using CellInput's transasction hash.
This change saves 32 bytes by not validating that the hash in CellInput
is indeed the hash of the context transaction; we maintain security
instead by simply using the calculated hash when deriving the current
transaction's hash. We will no longer provide feedback in the case where
an incorrect context transaction is sent, but it remains impossible to
fake the context transaction as the signature we produce will still be
for the transaction we presented to the user.