Tweak Bech32 encoder's handling of HRP
Though it's been fixed upstream,  this previously allowed non-ascii
characters to be part of the Human Readable Part (HRP).

It also disallowed capital characters, which is out of place in a
general-purpose Bech32 encoder and also necessary for Avalanche X-chain

Lastly, moved the encoder from null-delimited HRP to an explicit-length
HRP, which should avoid some memcpy's with the planned integration.

Explicit length also allowed a length check to be moved ahead of the
loop that previously did three things: validate the HRP, compute a
checksum state, and find the length of the HRP.  Now it only does the
first two.